Accomodate new SCAPI envelope
Right implementation doesn't do anything useful with an additional
envelope data, but it may be usefule for future extensions, like display
order (#1) or authentication.

Simple `id` counter may be removed or changed if needed.

There are no clear rules on how to handle afformentioned ID. They have to be
specifed when there will be a good use case for it.