Workaround for broken syntax highlighting
Syntax highlighting for complex xhtml files as of tooday is broken for vim. If
tags `<script/>` is encountered the the rest of file it highlighted as a single
string, but combination `<script></script>` is always replaced by
`xmllint --format` (which I always use). That's why I've added space, so
xmlling will leave tag as is and vim will be happy 😃.